+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## 0.0.3 - 2019-03-10
 
 ### Added
  - Onboarding views
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - README.md initialized
 - License added
 
-[Unreleased]: https://github.com/ngunyimacharia/qetela-client/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/ngunyimacharia/qetela-client/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/ngunyimacharia/qetela-client/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/ngunyimacharia/qetela-client/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/ngunyimacharia/qetela-client/compare/v0.0.0...v0.0.1